Add files via upload
A Morse code translator is a tool that converts text to Morse code and vice versa. Morse code is a method of encoding text characters as sequences of dots and dashes, which can be transmitted via sound, light, or visual signals. In a Morse code translator, each letter and number is mapped to a specific sequence of dots and dashes.